fix: made changes to information list file
