+++ packages/appstore/src/components/ready-to-update-wallets/ready-to-update-wallets.scss
@@ -36,6 +36,7 @@
 
     &--text {
         margin: 2.4rem 0;
+        gap: 0.8rem;
     }
 
     &--info-section {
